feat!: {iter->counts}::SiteCounts::try_new
# Conflicts:
#	src/counts.rs
