+++ src/stats.rs
@@ -1,6 +1,5 @@
-use crate::iter::SiteCounts;
 use crate::util::UnorderedPair;
-use crate::{Count, MultiPopulationCounts, MultiSiteCounts, PopgenError, PopgenResult};
+use crate::{Count, MultiPopulationCounts, MultiSiteCounts, PopgenError, PopgenResult, SiteCounts};
 use std::cmp::max;
 use std::collections::HashMap;
